pyproject.toml
[project]
name = "nextdns-rewrites"
version = "0.1.0"
description = "Manage NextDNS rewrites."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nextdns>=4.1.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
